refactor: relocate BlogList logic from pages to reusable component
Moved the blog listing functionality from the pages directory to a new reusable BlogList component under components. This improves code structure, promotes reusability, and aligns with best practices for component-driven development.
